Remove unnecessary unsigned casts
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1205 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_compile.c
@@ -3025,8 +3025,8 @@ switch(ptype)
   p = PRIV(ucd_caseless_sets) + prop->caseset;
   for (;;)
     {
-    if ((unsigned int)c < *p) return !negated;
-    if ((unsigned int)c == *p++) return negated;
+    if (c < *p) return !negated;
+    if (c == *p++) return negated;
     }
   break;  /* Control never reaches here */
 #endif
